Reassorted classes and renamed packages - (Task #410)
Reassorted classes in
	de.dlr.sc.virsat.excel
renamed packages in
	de.dlr.sc.virsat.model.extension.funcelectrial.test
	de.dlr.sc.virsat.model.extension.statemachines

---
Task #410: Improve package structure of excel plugins
