+++ de.dlr.sc.virsat.excel.test/src/de/dlr/sc/virsat/excel/ExcelExporterTest.java
@@ -21,6 +21,9 @@ import org.eclipse.emf.ecore.EObject;
 import org.junit.Before;
 import org.junit.Test;
 
+import de.dlr.sc.virsat.excel.exporter.ExcelExporter;
+import de.dlr.sc.virsat.excel.exporter.IExport;
+
 /**
  * This class test the ExcelExporter class.
  * @author muel_s8
