+++ de.dlr.sc.virsat.excel.test/src/de/dlr/sc/virsat/excel/ExcelImporterTest.java
@@ -27,7 +27,9 @@ import org.eclipse.emf.ecore.EObject;
 import org.junit.Before;
 import org.junit.Test;
 
+import de.dlr.sc.virsat.excel.fault.Fault;
 import de.dlr.sc.virsat.excel.importer.ExcelImporter;
+import de.dlr.sc.virsat.excel.importer.IImport;
 import de.dlr.sc.virsat.model.dvlm.DVLMFactory;
 import de.dlr.sc.virsat.model.dvlm.Repository;
 import de.dlr.sc.virsat.model.dvlm.structural.StructuralFactory;
