+++ de.dlr.sc.virsat.excel.ui/src/de/dlr/sc/virsat/excel/ui/wizards/ImportWizard.java
@@ -30,8 +30,8 @@ import org.eclipse.swt.widgets.Display;
 import org.eclipse.ui.INewWizard;
 import org.eclipse.ui.IWorkbench;
 
-import de.dlr.sc.virsat.excel.Fault;
 import de.dlr.sc.virsat.excel.commands.ImportCommand;
+import de.dlr.sc.virsat.excel.fault.Fault;
 import de.dlr.sc.virsat.excel.importer.ExcelImporter;
 import de.dlr.sc.virsat.model.dvlm.provider.DVLMEditPlugin;
 import de.dlr.sc.virsat.project.editingDomain.VirSatEditingDomainRegistry;
