+++ de.dlr.sc.virsat.excel/src/de/dlr/sc/virsat/excel/exporter/IExport.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.excel;
+package de.dlr.sc.virsat.excel.exporter;
 
 import org.eclipse.emf.ecore.EObject;
