+++ de.dlr.sc.virsat.excel/src/de/dlr/sc/virsat/excel/importer/ExcelImportHelper.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.excel;
+package de.dlr.sc.virsat.excel.importer;
 
 import java.util.List;
 import java.util.Objects;
